refactor: Move query builder interfaces to types file and add test step in CI
